+++ examples/next-simple/global.d.ts
@@ -1,4 +1,0 @@
-export declare module '*.svg' {
-  const content: string
-  export default content
-}