+++ examples/next-simple/src/constants.tsx
@@ -1,1 +1,0 @@
-export const APP_MAX_WIDTH = 1140