+++ examples/next-simple/src/pages/home/index.tsx
@@ -1,5 +1,0 @@
-import { Button } from '@ultraviolet/ui'
-
-const Home = () => <Button>Hello there!</Button>
-
-export default Home