+++ packages/ui/src/components/Checkbox/__stories__/Template.stories.tsx
@@ -1,11 +1,26 @@
 import type { StoryFn } from '@storybook/react'
+import { useState } from 'react'
 import { Checkbox } from '..'
 
 export const Template: StoryFn<typeof Checkbox> = ({
   'aria-label': ariaLabel,
   ...args
-}) => (
-  <Checkbox onChange={console.log} {...args}>
-    Beautiful checkbox
-  </Checkbox>
-)
+}) => {
+  const [checked, setChecked] = useState(false)
+
+  return (
+    <Checkbox
+      checked={checked}
+      onChange={e => setChecked(e.target.checked)}
+      {...args}
+    >
+      Beautiful checkbox
+    </Checkbox>
+  )
+}
+
+Template.args = {
+  disabled: false,
+  name: 'basic',
+  value: 'label-1',
+}