+++ packages/ui/src/components/MenuV2/constants.ts
@@ -1,0 +1,5 @@
+export const SIZES = {
+  small: '11.25rem',
+  medium: '17.5rem',
+  large: '23.75rem',
+}