+++ packages/ui/src/components/Text/index.tsx
@@ -1,7 +1,6 @@
 import type { Theme } from '@emotion/react'
 import styled from '@emotion/styled'
-import type React from 'react'
-import type { ElementType, ReactNode } from 'react'
+import type { CSSProperties, ElementType, ReactNode } from 'react'
 import { useRef } from 'react'
 import recursivelyGetChildrenString from '../../helpers/recursivelyGetChildrenString'
 import { useIsOverflowing } from '../../hooks/useIsOverflowing'
@@ -18,7 +17,8 @@ const PROMINENCES = {
 }
 
 type ProminenceProps = keyof typeof PROMINENCES
-type PlacementProps = React.CSSProperties['textAlign']
+type PlacementProps = CSSProperties['textAlign']
+type WhiteSpaceProps = CSSProperties['whiteSpace']
 type TextVariant = keyof typeof typography
 export const textVariants = Object.keys(typography) as TextVariant[]
 
@@ -35,6 +35,8 @@ const generateStyles = ({
   disabled,
   italic,
   underline,
+  whiteSpace,
+  strikeThrough,
 }: {
   placement?: PlacementProps
   prominence: ProminenceProps
@@ -45,6 +47,8 @@ const generateStyles = ({
   disabled: boolean
   italic: boolean
   underline: boolean
+  whiteSpace?: WhiteSpaceProps
+  strikeThrough?: boolean
 }): string => {
   // stronger is available only for neutral sentiment
   const definedProminence =
@@ -77,6 +81,7 @@ const generateStyles = ({
     text-transform: ${theme.typography[variant].textCase};
     text-decoration: ${theme.typography[variant].textDecoration};
     ${placement ? ` text-align: ${placement};` : ''}
+    ${whiteSpace ? `white-space: ${whiteSpace};` : ''}
 
     ${
       oneLine
@@ -87,6 +92,7 @@ const generateStyles = ({
     }
     ${italic ? `font-style: italic;` : ''}
     ${underline ? `text-decoration: underline;` : ''}
+    ${strikeThrough ? `text-decoration: line-through;` : ''}
   `
 }
 
@@ -110,6 +116,8 @@ type TextProps = {
   dir?: 'ltr' | 'rtl' | 'auto'
   htmlFor?: string
   'data-testid'?: string
+  strikeThrough?: boolean
+  whiteSpace?: WhiteSpaceProps
 }
 
 const StyledText = styled('div', {
@@ -124,6 +132,8 @@ const StyledText = styled('div', {
       'disabled',
       'italic',
       'underline',
+      'strikeThrough',
+      'whiteSpace',
     ].includes(prop),
 })<{
   placement?: PlacementProps
@@ -135,6 +145,8 @@ const StyledText = styled('div', {
   italic: boolean
   underline: boolean
   htmlFor?: string
+  strikeThrough?: boolean
+  whiteSpace?: WhiteSpaceProps
 }>(generateStyles)
 
 /**
@@ -153,8 +165,10 @@ export const Text = ({
   disabled = false,
   italic = false,
   underline = false,
+  strikeThrough = false,
   id,
   dir,
+  whiteSpace,
   htmlFor,
   'data-testid': dataTestId,
 }: TextProps) => {
@@ -178,10 +192,12 @@ export const Text = ({
         disabled={disabled}
         italic={italic}
         underline={underline}
+        strikeThrough={strikeThrough}
         id={id}
         dir={dir}
         htmlFor={htmlFor}
         data-testid={dataTestId}
+        whiteSpace={whiteSpace}
       >
         {children}
       </StyledText>